% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkcolnames.R
\name{checkcolnames}
\alias{checkcolnames}
\title{checkcolnames}
\usage{
checkcolnames(data, tab, soilcarbon_template)
}
\arguments{
\item{data}{soilcarbon data}

\item{tab}{specific tab of dataset , ie. dataset_name}

\item{soilcarbon_template}{template soilcarbon file to compare}
}
\description{
Check the column names of different tabs in the dataset to make sure they match, used in dataQC
}
