% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkdata.R
\name{checkdata}
\alias{checkdata}
\title{checkdata}
\usage{
checkdata(data)
}
\arguments{
\item{data}{directory to data file}
}
\description{
Check the imported soil carbon dataset for formatting and entry errors
}

