% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparenames.R
\name{comparenames}
\alias{comparenames}
\title{comparenames}
\usage{
comparenames(names)
}
\arguments{
\item{data}{directory to data file}

\item{name}{specific heirarchial name, ie. dataset_name}
}
\description{
Compares the names at names between hierarchical levels, used in dataQC
}

