% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileAsList.R
\name{compileAsList}
\alias{compileAsList}
\title{compileAsList}
\usage{
compileAsList(dataset_directory)
}
\arguments{
\item{dataset_directory}{directory where compeleted and QC passed soilcarbon datasets are stored}
}
\description{
Compiles all entries (xlsx files) and creates a version of the database as a list
}
