% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileDatabase.R
\name{compileDatabase}
\alias{compileDatabase}
\title{compileDatabase}
\usage{
compileDatabase(dataset_directory)
}
\arguments{
\item{dataset_directory}{directory where compeleted and QC passed soilcarbon datasets are stored}
}
\description{
adds dataset to soilcarbon database
}
