% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.Yujie.R
\name{convert.Yuije}
\alias{convert.Yuije}
\title{convert.Yujie}
\usage{
convert.Yuije(Yujie_file)
}
\arguments{
\item{Yujie_file}{directory to Yuije data file}
}
\description{
This function converts data compiled by Yuije into format compatable with the soilcarbon template
}
