% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataQC.R
\name{dataQC}
\alias{dataQC}
\title{dataQC}
\usage{
dataQC(data, tabs = c("metadata", "site", "profile", "layer", "fraction"),
  writeQCreport = F, outfile = NULL)
}
\arguments{
\item{data}{directory to data file}

\item{tabs}{the tabs that will be checked, default is all}

\item{writeQCreport}{if TRUE, a text report of the QC output will be written to the outfile. Default is FALSE}

\item{outfile}{filename of the output file if writeQCreport=TRUE. Default is NULL, and the outfile will be written to the directory where the dataset is stored, and named by the dataset being checked.}
}
\description{
Check the imported soil carbon dataset for formatting and entry errors
}
