% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{flatten}
\alias{flatten}
\title{flatten}
\usage{
flatten(soilcarbon_data)
}
\arguments{
\item{soilcarbon_data}{soilcarbon data oject}
}
\description{
converts heierachically strucutred soilcarbon data into flat data.frame
}
