% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getnames.R
\name{getnames}
\alias{getnames}
\title{getnames}
\usage{
getnames(data, name)
}
\arguments{
\item{data}{directory to data file}

\item{name}{specific heirarchial name, ie. dataset_name}
}
\description{
Extracts the names at names between hierarchical levels, used to compare names between levels
}
