% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg.R
\name{ghg}
\alias{ghg}
\title{Green House Gas Conversion}
\usage{
ghg(element, kg)
}
\arguments{
\item{This}{function will take inputs of either kg of Carbon, Methane, or Nitrous Oxide}
}
\value{
The function will return the conversion to kg of CO2e
}
\description{
Green House Gas Conversion
}
\author{
Tyler D. Cobian
}
