% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importdata.R
\name{importdata}
\alias{importdata}
\title{importdata}
\usage{
importdata(file)
}
\arguments{
\item{file}{directory to data file}
}
\description{
This function imports data from xlsx format matching the standard soil carbon data template
}
\examples{

importdata(file="~/path/to/file/file.xlsx")
}

