% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchnumeric.R
\name{matchnumeric}
\alias{matchnumeric}
\title{matchnumeric}
\usage{
matchnumeric(var_data, var_name, tab, error)
}
\arguments{
\item{var_data}{actual data values for a specific variable}

\item{var_name}{name of the variable being matched}

\item{tab}{name of the tab that the variable is found in}

\item{error}{error counter}
}
\description{
Check that values are numeric which are supposed to be
}
