% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchvocab.R
\name{matchvocab}
\alias{matchvocab}
\title{matchvocab}
\usage{
matchvocab(var_data, var_vocab, var_name, tab, error)
}
\arguments{
\item{var_data}{actual data values for a specific variable}

\item{var_vocab}{controlled vocab for a specific variable}

\item{var_name}{name of the variable being matched}

\item{tab}{name of the tab that the variable is found in}

\item{error}{error counter}
}
\description{
Match values in variable to controlled vocab
}
\examples{
matchvocab(var_data = c("A", "B"), var_vocab=c("A", "B", "C"), tab="site", error=0)
}
