% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.soilcarbon.R
\name{read.soilcarbon}
\alias{read.soilcarbon}
\title{read.soilcarbon}
\usage{
read.soilcarbon(file, format = "original", template = F)
}
\arguments{
\item{file}{directory to data file}

\item{format}{default is "original." Also accepted, "MPI" which modifies the read function for compatability with the MPI template V9.}

\item{template}{set to TRUE if reading in a template file}
}
\description{
This function imports data from xlsx format matching the standard soil carbon data template
}
