% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveYujieXlsx.R
\name{saveYujieXlsx}
\alias{saveYujieXlsx}
\title{saveYujieXlsx}
\usage{
saveYujieXlsx(directory)
}
\arguments{
\item{directory}{directory to save files}
}
\description{
This function saves Yujie database as individual .xlsx files matching the template
}
