% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_erosion.R
\name{soil_erosion}
\alias{soil_erosion}
\title{Soil Erosion Rates}
\usage{
soil_erosion(area, time)
}
\arguments{
\item{This}{function will take inputs of land are in units of hecatres and a unit of time in year}
}
\value{
This function will return the amount of top soil lost on kg after the specified time period
}
\description{
Soil Erosion Rates
}
\references{

}
\author{
Tyler D. Cobian
}
