% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_erosion_data.R
\docType{data}
\name{soil_erosion_data}
\alias{soil_erosion_data}
\title{Example data set of soil erosion rates}
\format{An object of class \code{data.frame} with 100 rows and 3 columns.}
\usage{
soil_erosion_data
}
\description{
\describe(
\item(year){time elapsed in years}
\item(land_area){Hecatares of land accounted for}
\item(erosion.year){kg of soil that have been lost since year 1})
}
\keyword{datasets}
