% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_properties_data.R
\docType{data}
\name{soil_properties}
\alias{soil_properties}
\title{Data on different averge soil properties}
\format{An object of class \code{matrix} with 100 rows and 3 columns.}
\usage{
soil_properties
}
\description{
Data accounts for Percent Soil Carbon, Bulk Density, and Percent Clay Content
}
\details{
\describe(
 \item(bulk_density){Bulk Density of soil in mg/kg})
}
\keyword{datasets}
