% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilcarbon_data.R
\docType{data}
\name{soilcarbon_data}
\alias{soilcarbon_data}
\title{Data compiled for the soilcarbon dataset}
\format{A data frame with 4115 rows and 42 variables:
\describe{
  \item{dataset_name}{name of the dataset that observations were collected from}
  \item{site_name}{name of the site where data was gathered}
  ...
}}
\source{
\url{https://powellcenter.usgs.gov/view-project/55d4bffce4b0518e354695aa}
}
\usage{
soilcarbon_data
}
\description{
A dataset containing soil carbon parameters collected by the powell center soil carbon working group
}
\keyword{datasets}

