% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabfromHe.R
\name{tabfromHe}
\alias{tabfromHe}
\title{tabfromHe}
\usage{
tabfromHe(dir, tab, Author)
}
\arguments{
\item{dir}{directory where Yujie He database and Master_template files are stored in your local computer}

\item{tab}{target tab to be created. Only site, profile, layer, fraction are valid}

\item{Author}{Author of data to be extracted}
}
\value{
a data.frame with columns matching those of the template and additional columns appended at the end
}
\description{
Creates a tab from Yujie He's database using the format of the Master_template
}
