% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabReport.R
\name{tabReport}
\alias{tabReport}
\title{tabReport}
\usage{
tabReport(dataset_directory, tab, cols)
}
\arguments{
\item{dataset_directory}{directory where compeleted and QC passed soilcarbon datasets are stored}

\item{tab}{tab selected to generate a report}

\item{cols}{set of columns to be selected from the specified tab}
}
\value{
a data.frame
}
\description{
Compiles all entries (xlsx files) and creates report from a specific tab and a specific set of columns
}
\examples{
# tabReport(dataset_directory = "data_files/", tab="metadata",cols=c("dataset_name","curator_name", "curator_email"))

}
\author{
Carlos A. Sierra
}
