% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checknames.R
\name{checknames}
\alias{checknames}
\title{checknames}
\usage{
checknames(data, name, tabs)
}
\arguments{
\item{data}{soilcarbon data}

\item{name}{specific heirarchial name, ie. dataset_name}

\item{tabs}{the tabs that will be checked, default is all}
}
\description{
Check the names of different levels in the dataset to make sure they match, used in dataQC
}
