% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkvalues.R
\name{checkvalues}
\alias{checkvalues}
\title{checkvalues}
\usage{
checkvalues(data, tab)
}
\arguments{
\item{data}{soilcarbon data}

\item{tab}{specific tab of dataset , ie. dataset_name}
}
\description{
Check certain column values to match what is expected, used in dataQC
}
