% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparenames.R
\name{comparenames}
\alias{comparenames}
\title{comparenames}
\usage{
comparenames(names)
}
\arguments{
\item{names}{names oject, outout from getnames()}
}
\description{
Compares the names at names between hierarchical levels, used in dataQC
}
