% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.soilcarbon.R
\name{read.soilcarbon}
\alias{read.soilcarbon}
\title{read.soilcarbon}
\usage{
read.soilcarbon(file)
}
\arguments{
\item{file}{directory to data file}
}
\description{
This function imports data from xlsx format matching the standard soil carbon data template
}
\examples{
datafile<-system.file("extdata", "UNALTERED_TEMPLATE_v120616.xlsx", package = "soilcarbon")
read.soilcarbon(file=datafile)
}

