% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilcarbon_database.R
\docType{data}
\name{soilcarbon_database}
\alias{soilcarbon_database}
\title{Data compiled for the soilcarbon dataset}
\format{A data frame with 4155 rows and 42 variables:
\describe{
\item{dataset_name}{}
\item{site_name}{}
\item{profile_name}{}
\item{lat}{}
\item{long}{}
\item{parent_material}{}
\item{slope}{}
\item{slope_shape}{}
\item{aspect}{}
\item{veg_note}{}
\item{country}{}
\item{state_province}{}
\item{mat}{}
\item{map}{}
\item{elevation}{}
\item{observation_date}{}
\item{soil_taxon}{}
\item{layer_name}{}
\item{layer_top}{}
\item{layer_bot}{}
\item{hzn}{}
\item{rc_year}{}
\item{X13c}{}
\item{X14c}{}
\item{X14c_sigma}{}
\item{bd_tot}{}
\item{bet_surface_area}{}
\item{ph_h2o}{}
\item{c_tot}{}
\item{n_tot}{}
\item{c_to_n}{}
\item{sand_tot_psa}{}
\item{silt_tot_psa}{}
\item{clay_tot_psa}{}
\item{cat_exch}{}
\item{fe_dith}{}
\item{fe_ox}{}
\item{fe_py}{}
\item{al_py}{}
\item{al_dith}{}
\item{al_ox}{}
\item{smect_vermic}{}
\item{doi_number}{}
\item{curator_name}{}
\item{curator_organization}{}
\item{curator_email}{}
\item{modification_date}{}
\item{contact_name}{}
\item{contact_email}{}
\item{contact_orcid_id}{}
\item{bibliographical.reference}{}
\item{datum}{}
\item{site_note}{}
\item{climate_cat}{}
\item{aspect_deg}{}
\item{drainage_class}{}
\item{depth_water}{}
\item{parent_chem}{}
\item{ecoregion}{}
\item{land_cover}{}
\item{s_soc}{}
\item{profile_note}{}
\item{profile_comp}{}
\item{veg_note_profile}{}
\item{X2d_position}{}
\item{soil_series}{}
\item{bedrock_depth}{}
\item{thaw_depth_profile}{}
\item{p_soc}{}
\item{p_soc_sigma}{}
\item{p_soc_depth}{}
\item{layer_comp}{}
\item{layer_note}{}
\item{color}{}
\item{burn_ev}{}
\item{bd_samp}{}
\item{bd_notes}{}
\item{ph_cacl}{}
\item{ph_other}{}
\item{caco3}{}
\item{coarse_tot}{}
\item{coarse_size_thresh}{}
\item{texture_class}{}
\item{base_sum}{}
\item{cec_sum}{}
\item{ecec}{}
\item{bs}{}
\item{soc}{}
\item{soc_sigma}{}
\item{oc}{}
\item{loi}{}
\item{X15n}{}
\item{rc_lab}{}
\item{rc_lab_number}{}
\item{fraction_modern}{}
\item{fraction_modern_sigma}{}
\item{mbc_method}{}
\item{raw_mbc}{}
\item{reported_mbc}{}
\item{p_ext}{}
\item{p_units}{}
\item{p_method}{}
\item{si_py}{}
\item{c_py}{}
\item{py_notes}{}
\item{si_ox}{}
\item{c_ox}{}
\item{ox_notes}{}
\item{fe_hy}{}
\item{al_hy}{}
\item{si_hy}{}
\item{c_hy}{}
\item{hy_notes}{}
\item{si_dith}{}
\item{dith_notes}{}
\item{quartz}{}
\item{alkali_feldspar}{}
\item{plag_feldspar}{}
\item{mica_chlorite}{}
\item{amphibole}{}
\item{pyroxine}{}
\item{olivine}{}
\item{volc_glass}{}
\item{kaol_halloy}{}
\item{gibbsite}{}
\item{fe_oxides}{}
\item{imog_alloph}{}
\item{ferrihydrite}{}
\item{fraction_name}{}
\item{f_scheme}{}
\item{f_property}{}
\item{f_scheme_units}{}
\item{f_lower}{}
\item{f_upper}{}
\item{f_scheme_conc}{}
\item{f_seq}{}
\item{f_seq_order}{}
\item{f_comp}{}
\item{f_note}{}
\item{f_c_perc}{}
\item{f_mass_perc}{}
\item{f_c_tot}{}
\item{f_oc}{}
\item{f_n_tot}{}
\item{f_c_to_n}{}
\item{f_15n}{}
\item{f_13c}{}
\item{f_rc_lab}{}
\item{f_rc_lab_number}{}
\item{f_rc_year}{}
\item{f_14c}{}
\item{f_14c_sigma}{}
\item{f_fraction_modern}{}
\item{f_fraction_modern_sigma}{}
\item{f_quartz}{}
\item{f_alkali_feldspar}{}
\item{f_plag_feldspar}{}
\item{f_mica_chlorite}{}
\item{f_amphibole}{}
\item{f_pyroxine}{}
\item{f_olivine}{}
\item{f_volc_glass}{}
\item{f_kaol_halloy}{}
\item{f_smect_vermic}{}
\item{f_gibbsite}{}
\item{f_fe_oxides}{}
\item{f_imog_alloph}{}
\item{f_ferrihydrite}{}
  ...
}}
\source{
\url{https://powellcenter.usgs.gov/view-project/55d4bffce4b0518e354695aa}
}
\usage{
soilcarbon_database
}
\description{
A dataset containing soil carbon parameters collected by the powell center soil carbon working group
}
\keyword{datasets}
